Implement OpenAPI configurator interface and update OpenApiPlugin to support configurators
